poetry refactor
