+++ l2s/__init__.py
@@ -1,2 +1,2 @@
 #!/usr/bin/env python3
-from api import *
+from l2s.api import *
