+++ l2s/api.py
@@ -1,4 +1,5 @@
-import _simple_sim
+from l2s._simple_sim import simple_sim, map_neuron_type_to_int
+from cri_simulations import network
 
 class CRI_network:
 
@@ -12,6 +13,10 @@ class CRI_network:
         self.simpleSim = None
         self.target = target
         #we may need more attributes to track hardware implementation
+        if(self.target == 'CRI'):
+            print('Initilizing to run on hardware')
+            self.CRI = network(self.axons, self.connections, self.inputs, {}, self.config)
+            self.CRI.initalize_network()
 
     def update_synapse(self):
         pass
@@ -22,7 +27,7 @@ class CRI_network:
         """
         if (self.target == "simpleSim"):
             if(not self.simpleSim):
-                self.simpleSim = _simple_sim.simple_sim(_simple_sim.map_neuron_type_to_int(self.config['neuron_type']), self.config['global_neuron_params']['v_thr'], self.axons, self.connections, self.inputs)
+                self.simpleSim = simple_sim(map_neuron_type_to_int(self.config['neuron_type']), self.config['global_neuron_params']['v_thr'], self.axons, self.connections, self.inputs)
 
             self.simpleSim.free_run()
         else:
@@ -31,8 +36,11 @@ class CRI_network:
     def step(self,target="simpleSim"):
         if (self.target == "simpleSim"):
             if(not self.simpleSim):
-                _simple_sim.simple_sim(_simple_sim.map_neuron_type_to_int(self.config['neuron_type']), self.config['global_neuron_params']['v_thr'], self.axons, self.connections, self.inputs)
+                self.simpleSim = simple_sim(map_neuron_type_to_int(self.config['neuron_type']), self.config['global_neuron_params']['v_thr'], self.axons, self.connections, self.inputs)
 
-            _simple_sim.step_run()
+            self.simpleSim.step_run()
+        elif (self.target == "CRI"):
+            return self.CRI.run_step()
         else:
             raise Exception("Invalid Target")
+
